Add exact request budget preflight
Add a side-effect-free chat request preflight API that uses the production prompt preparation path, enforces context and output budgets, and exposes asynchronous JavaScript support without blocking the caller.


Copilot-Session: 6883972e-cc0b-4076-a3a5-86106e7a1133

+++ sdk_v2/cpp/src/inferencing/generative/genai_config.cc
@@ -119,7 +119,7 @@ GenAIConfig GenAIConfig::LoadFromFile(const std::string& path) {
     OnnxModel model;
 
     if (jm.contains("context_length") && jm["context_length"].is_number()) {
-      model.context_length = jm["context_length"].get<int>();
+      model.context_length = jm["context_length"].get<int64_t>();
     }
 
     if (jm.contains("type") && jm["type"].is_string()) {
@@ -173,7 +173,7 @@ GenAIConfig GenAIConfig::LoadFromFile(const std::string& path) {
     Search search;
 
     if (js.contains("max_length") && js["max_length"].is_number()) {
-      search.max_length = js["max_length"].get<int>();
+      search.max_length = js["max_length"].get<int64_t>();
     }
 
     config.search = std::move(search);
